ints: change some ints to int32
- this is to help reduce the overall size of some proto prs.

+++ privval/file.go
@@ -85,18 +85,18 @@ type FilePVLastSignState struct {
 // it returns true if the HRS matches the arguments and the SignBytes are not empty (indicating
 // we have already signed for this HRS, and can reuse the existing signature).
 // It panics if the HRS matches the arguments, there's a SignBytes, but no Signature.
-func (lss *FilePVLastSignState) CheckHRS(height int64, round int, step int8) (bool, error) {
+func (lss *FilePVLastSignState) CheckHRS(height int64, round int32, step int8) (bool, error) {
 
 	if lss.Height > height {
 		return false, fmt.Errorf("height regression. Got %v, last height %v", height, lss.Height)
 	}
 
 	if lss.Height == height {
-		if lss.Round > round {
+		if lss.Round > int(round) {
 			return false, fmt.Errorf("round regression at height %v. Got %v, last round %v", height, round, lss.Round)
 		}
 
-		if lss.Round == round {
+		if lss.Round == int(round) {
 			if lss.Step > step {
 				return false, fmt.Errorf(
 					"step regression at height %v round %v. Got %v, last step %v",
@@ -375,11 +375,11 @@ func (pv *FilePV) signProposal(chainID string, proposal *types.Proposal) error {
 }
 
 // Persist height/round/step and signature
-func (pv *FilePV) saveSigned(height int64, round int, step int8,
+func (pv *FilePV) saveSigned(height int64, round int32, step int8,
 	signBytes []byte, sig []byte) {
 
 	pv.LastSignState.Height = height
-	pv.LastSignState.Round = round
+	pv.LastSignState.Round = int(round)
 	pv.LastSignState.Step = step
 	pv.LastSignState.Signature = sig
 	pv.LastSignState.SignBytes = signBytes
